+++ src/main/java/org/openlmis/fulfillment/Application.java
@@ -35,7 +35,6 @@ import org.springframework.validation.beanvalidation.LocalValidatorFactoryBean;
 import org.springframework.web.servlet.LocaleResolver;
 import org.springframework.web.servlet.i18n.CookieLocaleResolver;
 
-import java.time.Clock;
 import java.util.Locale;
 
 @SpringBootApplication(scanBasePackages = "org.openlmis.fulfillment")
@@ -128,11 +127,6 @@ public class Application {
     };
 
     return strategy;
-  }
-
-  @Bean
-  public Clock clock() {
-    return Clock.systemDefaultZone();
   }
 
 }