+++ src/main/java/org/openlmis/fulfillment/repository/custom/OrderRepositoryCustom.java
@@ -18,7 +18,6 @@ package org.openlmis.fulfillment.repository.custom;
 import org.openlmis.fulfillment.domain.Order;
 import org.openlmis.fulfillment.domain.OrderStatus;
 
-import java.time.LocalDate;
 import java.util.List;
 import java.util.Set;
 import java.util.UUID;
@@ -26,8 +25,7 @@ import java.util.UUID;
 public interface OrderRepositoryCustom {
 
   List<Order> searchOrders(UUID supplyingFacility, UUID requestingFacility, UUID program,
-                           UUID processingPeriod, Set<OrderStatus> statuses, LocalDate startDate,
-                           LocalDate endDate);
+                           UUID processingPeriod, Set<OrderStatus> statuses);
 
   List<UUID> getRequestingFacilities(UUID supplyingFacility);