+++ src/main/java/org/openlmis/fulfillment/repository/custom/impl/OrderRepositoryImpl.java
@@ -26,12 +26,7 @@ import static org.springframework.util.CollectionUtils.isEmpty;
 import org.openlmis.fulfillment.domain.Order;
 import org.openlmis.fulfillment.domain.OrderStatus;
 import org.openlmis.fulfillment.repository.custom.OrderRepositoryCustom;
-import org.openlmis.fulfillment.util.DateHelper;
-import org.springframework.beans.factory.annotation.Autowired;
 
-import java.time.LocalDate;
-import java.time.LocalTime;
-import java.time.ZonedDateTime;
 import java.util.Collection;
 import java.util.List;
 import java.util.Set;
@@ -46,9 +41,6 @@ import javax.persistence.criteria.Root;
 
 public class OrderRepositoryImpl implements OrderRepositoryCustom {
 
-  @Autowired
-  private DateHelper dateHelper;
-
   @PersistenceContext
   private EntityManager entityManager;
 
@@ -60,14 +52,11 @@ public class OrderRepositoryImpl implements OrderRepositoryCustom {
    * @param program            program of searched Orders.
    * @param processingPeriod   UUID of processing period
    * @param statuses           order statuses.
-   * @param startDate          start date
-   * @param endDate            end date
    * @return List of Orders with matched parameters.
    */
   @Override
   public List<Order> searchOrders(UUID supplyingFacility, UUID requestingFacility,
-                                  UUID program, UUID processingPeriod, Set<OrderStatus> statuses,
-                                  LocalDate startDate, LocalDate endDate) {
+                                  UUID program, UUID processingPeriod, Set<OrderStatus> statuses) {
     CriteriaBuilder builder = entityManager.getCriteriaBuilder();
     CriteriaQuery<Order> query = builder.createQuery(Order.class);
     Root<Order> root = query.from(Order.class);
@@ -78,7 +67,6 @@ public class OrderRepositoryImpl implements OrderRepositoryCustom {
     predicate = isEqual(PROGRAM_ID, program, root, predicate, builder);
     predicate = isEqual(PROCESSING_PERIOD_ID, processingPeriod, root, predicate, builder);
     predicate = isOneOf(STATUS, statuses, root, predicate, builder);
-    predicate = isBetween(CREATED_DATE, startDate, endDate, root, predicate, builder);
 
     query.where(predicate);
     query.orderBy(builder.desc(root.get(CREATED_DATE)));
@@ -114,30 +102,6 @@ public class OrderRepositoryImpl implements OrderRepositoryCustom {
       }
 
       return builder.and(predicate, collectionPredicate);
-    }
-
-    return predicate;
-  }
-
-  private Predicate isBetween(String field, LocalDate start, LocalDate end,
-                              Root<Order> root, Predicate predicate, CriteriaBuilder builder) {
-    ZonedDateTime from = null;
-    ZonedDateTime to = null;
-
-    if (start != null) {
-      from = start.atStartOfDay(dateHelper.getZone());
-    }
-
-    if (end != null) {
-      to = ZonedDateTime.of(end, LocalTime.MAX, dateHelper.getZone());
-    }
-
-    if (from != null && to != null) {
-      return builder.and(predicate, builder.between(root.get(field), from, to));
-    } else if (from != null) {
-      return builder.and(predicate, builder.greaterThanOrEqualTo(root.get(field), from));
-    } else if (to != null) {
-      return builder.and(predicate, builder.lessThanOrEqualTo(root.get(field), to));
     }
 
     return predicate;